pyproject.toml
[project]
name = "vantrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.15",
    "fastembed>=0.7.1",
    "html5lib>=1.1",
    "llama-index>=0.12.52",
    "lxml>=6.0.0",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "scrapy>=2.13.3",
    "sentence-transformers>=5.0.0",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
]
